1. 푸시/알림 닉네임, 이름 변경 제목 제외 요청   ( 필요시 복사 붙여넣기로 사용한다고 함 ) 2. 조회 조건 기준 개발중 3. 조회 조건 변경중  1) 로그인현황  2) 가입/탈퇴통계  3) 결제현황  4) IOS인앱결제시도  5) 현금영수증 발행내역  6) 자원관리   - 시간대별   - 월간별   - 연간별   - 회원Data   - 달구매내역   - 달사용내역   - 교롼내역

+++ src/main/java/com/dalbit/money/vo/Mon_AutoChangeOutputVo.java
@@ -26,11 +26,11 @@ public class Mon_AutoChangeOutputVo {
     private int testCnt;
     private int testAutoCnt;
     private int testAutoByeol;
-    private int totalExchangeAmt;
-    private int maleExchangeAmt;
-    private int femaleExchangeAmt;
-    private int noneExchangeAmt;
-    private int testExchangeAmt;
+    private long totalExchangeAmt;
+    private long maleExchangeAmt;
+    private long femaleExchangeAmt;
+    private long noneExchangeAmt;
+    private long testExchangeAmt;
     private int dateDiff;
 
     // output
